refactor: move svg-payload-scan into src/sanitize/
