+++ src/css-sanitizer.ts
@@ -30,7 +30,7 @@ import valueParser, {
     Node as ValueNode,
     FunctionNode
 } from 'postcss-value-parser';
-import { isSafeImageDataUri } from './svg-payload-scan';
+import { isSafeImageDataUri } from './sanitize/svg-payload-scan';
 import { recordRemoval } from './diagnostics/diagnostics-sink';
 
 const ALLOWED_AT_RULES = new Set<string>([
