+++ src/sanitize/svg-payload-scan.ts
